Add action replies and more call skeleton
This commit has a lot of changes that are not related but are
mixed together. More classes have been added for the video
conferencing example, and a new mechanism has been added which
allows actions to specify that they want a reply from the user
(e.g, "are you sure?").

 - Add Call.java, Contact.java and CallState.java
 - Add CallContextActionMap.java
 - OverworldTest.java: update open tests
 - OpenObject.java: add example of confirmation
 - VoiceProcess.java: add action reply mechanism